[REFACTOR] Custom Calendar View 수정
- ViewModel / ViewController 분리
- Swipe Gesture 추가
