--- iOS/AreUDone/AreUDone/CalendarScene/Model/Day.swift
@@ -1,21 +1,0 @@
-//
-//  Day.swift
-//  AreUDone
-//
-//  Created by a1111 on 2020/11/24.
-//
-
-import Foundation
-
-struct Day {
-  // 1
-  let date: Date
-  // 2
-  let number: String
-  // 3
-  let isSelected: Bool
-  // 4
-  let isWithinDisplayedMonth: Bool
-}
-
-
